### 2016-05-17 #21 @whedon
Submission: pyuca: a Python implementation of the Unicode Collation Algorithm
**Submitting author:** @jtauber (<a href="http://orcid.org/0000-0001-6534-8866">James Tauber</a>)
**Repository:** <a href="https://github.com/jtauber/pyuca" target ="_blank">https://github.com/jtauber/pyuca</a>
**Version:** v1.1.2
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.51550" target ="_blank">10.5281/zenodo.51622</a>

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.51550">software archive</a> resolve?

- [x] **Performance:** Have the performance claims of the software been confirmed?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

Compiled paper PDF: [10.21105.joss.00021.pdf](https://github.com/openjournals/joss-reviews/files/268914/10.21105.joss.00021.pdf)
