### 2016-05-17 #21 @luizirber
- There are no performance claims, so should I will just check the box?
- Although `test.py` have examples, the `usage` in the README could include an example `words` list (like `["cafe", "caff", "café"]` from [this test](https://github.com/jtauber/pyuca/blob/e048ae43ccc89d33e3d31777701c0e4bdabe8d47/test.py#L17))
- The author affiliation is missing in paper.md
- There are no community guidelines (Add a CONTRIBUTING file?)

Just some minor corrections, looks very good otherwise! @jtauber @arfon
