### 2016-05-18 #21 @jtauber
I'll add the `["cafe", "caff", "café"]` example to README and add a CONTRIBUTING file.

As to affiliation, I normally leave it blank as my scholarly work is independent (on academia.edu I'm under independent.academia.edu). I *could* put my company but it's not really relevant to this work.
