### 2016-05-18 6772899
more comphensive usage example
