### 2016-05-18 aeed204
added contributing guidelines
