+++ CONTRIBUTING.md
@@ -1,0 +1,7 @@
+Pull requests are very welcome although for sizable tasks it may be worth
+discussing via a GitHub issue before undertaking too much work.
+
+All tests, including flake8 should pass and code coverage should stay at 100%.
+
+If you have any problems, questions or suggestions, please file an issue on
+GitHub.
