### 2016-05-18 7c27976
added section to README on contacting the developer
