### 2016-05-18 #21 @jtauber
@luizirber @arfon

I've updated the `README.md` and added a `CONTRIBUTING.md`.

One question (besides the outstanding affiliation question): do I need to re-release and mint a new DOI to include these changes?
