### 2016-05-18 #21 @luizirber
@jtauber I left a [comment](https://github.com/jtauber/pyuca/commit/67728992e1de1d9370804da8f005a1b8c82b25a1#commitcomment-17517151) in the commit, I'll let @arfon answer about the affiliation and re-relase and new DOI questions.
