### 2016-05-18 #21 @arfon
> One question (besides the outstanding affiliation question): do I need to re-release and mint a new DOI to include these changes?

Ideally yes please!
