### 2016-05-18 871cf5d
fixed usage example to be standalone
previously it was just copy-pasted from unit test and didn't work without that context
