### 2016-05-18 #21 @jtauber
@luizirber I've addressed your comment. Thank you!
@arfon no problem doing another release. Did you want me to do anything about affiliation first?
