### 2016-05-18 #21 @arfon
> @arfon no problem doing another release. Did you want me to do anything about affiliation first?

Leaving it blank is fine.
