### 2016-05-18 #21 @jtauber
@luizirber @arfon re-released with new DOI: 10.5281/zenodo.51622
