### 2016-05-18 #21 @jtauber
Note the version has changed to 1.1.2 too.
