### 2016-05-18 #21 @arfon
@jtauber http://dx.doi.org/10.21105/joss.00021 🎉
